requirements.txt
# ── Core data & numerics ──────────────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1

# ── Machine learning ──────────────────────────────────────────────────────────
scikit-learn==1.5.0
joblib==1.4.2

# ── Deep learning (TensorFlow — CPU default; swap for tensorflow-gpu if needed)
tensorflow==2.16.1

# ── Reinforcement learning ────────────────────────────────────────────────────
stable-baselines3==2.3.2
gymnasium==0.29.1
shimmy==1.3.0          # gymnasium compatibility shim for stable-baselines3

# ── Data collection & APIs ────────────────────────────────────────────────────
requests==2.32.3
httpx==0.27.0
python-dotenv==1.0.1   # loads .env for secret API keys

# ── Config & utilities ────────────────────────────────────────────────────────
pyyaml==6.0.1
click==8.1.7           # CLI argument parsing in main.py
rich==13.7.1           # pretty terminal output & progress bars
tqdm==4.66.4

# ── Visualisation ─────────────────────────────────────────────────────────────
matplotlib==3.9.0
seaborn==0.13.2
plotly==5.22.0

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit==1.35.0
streamlit-option-menu==0.3.12

# ── Reporting ─────────────────────────────────────────────────────────────────
fpdf2==2.7.9           # PDF report export from dashboard
openpyxl==3.1.4        # Excel export for simulation results

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-cov==5.0.0
